Move common code for drawing about box and info viewer to a base class, remove the last of the info viewer logic from the base menu class.
