+++ src/frontend/mame/ui/about.h
@@ -12,7 +12,8 @@
 
 #pragma once
 
-#include "ui/menu.h"
+#include "ui/text.h"
+#include "ui/textbox.h"
 
 #include <optional>
 #include <string>
@@ -21,7 +22,7 @@
 
 namespace ui {
 
-class menu_about : public menu
+class menu_about : public menu_textbox
 {
 public:
 	menu_about(mame_ui_manager &mui, render_container &container);
@@ -30,13 +31,13 @@ public:
 protected:
 	virtual void custom_render(void *selectedref, float top, float bottom, float x, float y, float x2, float y2) override;
 
+	virtual float populate_text(std::optional<text_layout> &layout, float width) override;
+
 private:
-	virtual void draw(uint32_t flags) override;
 	virtual void populate(float &customtop, float &custombottom) override;
 	virtual void handle() override;
 
 	std::vector<std::string> const m_header;
-	std::optional<text_layout> m_layout;
 };
 
 } // namespace ui
